fix: move interfaces/ dir under the src/
