+++ src/components/help_portal.tsx
@@ -1,10 +1,10 @@
 import { Box, Text } from 'ink';
 import React, { FC, useEffect, useState } from 'react';
+import { Action, ActionablePage, pageActions } from '../../libs/notios-config/src/action_definitions';
 import {
   NotiosConfigActionKeymapping,
   NotiosConfigKeymappingRoot,
-} from '../../libs/notios-config/interfaces/notios-config';
-import { Action, ActionablePage, pageActions } from '../../libs/notios-config/src/action_definitions';
+} from '../../libs/notios-config/src/interfaces/notios-config';
 import { useNotiosConfigContext } from '../contexts/notios_config_context';
 import { useTermShapeContext } from '../contexts/term_shape_context';
 import useAction from '../hooks/use_action';
